Fix spill/fill response sampling
If the call request generates a spill, an additional cycle is required to
sample the response. In case of a ret request instead, the fill output can be
sampled the subsequent cycle. In either case, the problem of skipping request
sampling while the bypass is high is handled in a dedicated branch of
the capture() task.